% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_bridge.R
\name{java_ref_lambdas}
\alias{java_ref_lambdas}
\title{Reference Lambdas}
\usage{
java_ref_lambdas(handle)
}
\arguments{
\item{handle}{A \code{jobjRef} returned by \code{java_ref_run()$handle}.}
}
\value{
Numeric vector of feature weights.
}
\description{
Reference Lambdas
}
