% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R
\name{mock_occurrences_asym}
\alias{mock_occurrences_asym}
\title{Load Asymmetric Mock Occurrence Records (Phase B)}
\usage{
mock_occurrences_asym()
}
\value{
Data frame with \code{species}, \code{lon}, \code{lat}.
}
\description{
Load Asymmetric Mock Occurrence Records (Phase B)
}
